requirements.txt
opencv-python
numpy
pillow
diffusers
torch